requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
cffi=1.16.0=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
dss-python=0.15.7=pypi_0
dss-python-backend=0.14.5=pypi_0
fonttools=4.53.0=pypi_0
h5py=3.12.1=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_1
matplotlib=3.9.0=pypi_0
numpy=2.0.0=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py311haa95532_0
pycparser=2.22=pypi_0
pyparsing=3.1.2=pypi_0
python=3.11.9=he1021f5_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=69.5.1=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wheel=0.43.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
